+++ pytensor/compile/debug/monitormode.py
