+++ pytensor/compile/debug/nanguardmode.py
