+++ pytensor/compile/debug/profiling.py
