+++ scripts/mypy-failing.txt
@@ -1,6 +1,6 @@
 pytensor/compile/aliasing.py
 pytensor/compile/builders.py
-pytensor/compile/debugmode.py
+pytensor/compile/debug/debugmode.py
 pytensor/compile/executor.py
 pytensor/compile/maker.py
 pytensor/compile/rebuild.py
