+++ tests/compile/debug/test_monitormode.py
@@ -3,7 +3,7 @@ from io import StringIO
 
 import numpy as np
 
-from pytensor.compile import MonitorMode
+from pytensor.compile.debug.monitormode import MonitorMode
 from pytensor.compile.maker import function
 from pytensor.printing import debugprint
 from pytensor.tensor import log, outer
