+++ tests/compile/debug/test_nanguardmode.py
@@ -9,8 +9,8 @@ import pytest
 
 import pytensor.tensor as pt
 from pytensor.compile import shared
+from pytensor.compile.debug.nanguardmode import NanGuardMode
 from pytensor.compile.maker import function
-from pytensor.compile.nanguardmode import NanGuardMode
 from pytensor.configdefaults import config
 from pytensor.tensor.math import dot
 from pytensor.tensor.type import matrix, tensor3
