+++ tests/tensor/test_gc.py
@@ -65,7 +65,7 @@ def test_gc_never_pickles_temporaries():
         assert a(g) == a(g)  # some sanity checks on the pickling mechanism
 
         def b(fn):
-            return len(pickle.dumps(pytensor.compile.maker._pickle_Function(fn)))
+            return len(pickle.dumps(pytensor.compile.executor._pickle_Function(fn)))
 
         assert b(f) == b(f)  # some sanity checks on the pickling mechanism
