+++ tests/test_printing.py
@@ -11,6 +11,7 @@ import pytest
 
 import pytensor
 from pytensor import config
+from pytensor.compile.debug.profiling import ProfileStats
 from pytensor.compile.mode import get_mode
 from pytensor.compile.ops import deep_copy_op
 from pytensor.printing import (
@@ -109,7 +110,7 @@ def test_pydotprint_long_name():
 )
 def test_pydotprint_profile():
     A = matrix()
-    prof = pytensor.compile.ProfileStats(atexit_print=False, gpu_checks=False)
+    prof = ProfileStats(atexit_print=False, gpu_checks=False)
     f = pytensor.function([A], A + 1, profile=prof)
     pydotprint(f, print_output_file=False)
     f([[1]])
